Moved pages to schedules folder
